CH2 L5: Start

+++ functions/write_file_content.py
@@ -7,13 +7,20 @@ def write_file(working_directory, file_path, content):
     if not abs_file_path.startswith(abs_working_directory):
         return f'Error: Cannot write to "{file_path}" as it is outside the permitted working directory'
     
+    if not os.path.exists(abs_file_path):
+        try:
+            os.makedirs(os.path.dirname(abs_file_path), exist_ok=True)
+        except Exception as e:
+            return f'Error: creating directory: {e}'
+        
+    if os.path.exists(abs_file_path) and os.path.isdir(abs_file_path):
+        return f'Error: "{file_path}" is a directory, not a file'
+    
     try:
-        if not os.path.exists(abs_file_path):
-            os.makedirs(file_path, exist_ok=True)
         with open(abs_file_path, 'w') as f:
             f.write(content)
         return f'Successfully wrote to "{file_path}" ({len(content)} characters written)'
     except Exception as e:
-        return f'Error: {e}'
+        return f'Error: writing to file: {e}'
